refatorando e passando as logicas de negocio das telas para um customhook
