+++ src/components/Toast/ToastContainer/index.js
@@ -1,28 +1,11 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 
 import { Container } from './styles';
 import ToastMessage from '../ToastMessage';
-import { ToastEventManager } from '../../../utils/toast';
+import useToastContainer from './useToastContainer';
 
 function ToastContainer() {
-  const [messages, setMessages] = useState([]);
-
-  useEffect(() => {
-    function handleAddToast({ type, text }) {
-      setMessages((prevState) => [...prevState, { id: Math.random(), type, text }]);
-    }
-
-    ToastEventManager.on('addtoast', handleAddToast);
-
-    return () => {
-      ToastEventManager.removeListener('addtoast', handleAddToast);
-    };
-  }, []);
-
-  const handleRemoveMessage = useCallback((id) => {
-    setMessages((prevState) => prevState.filter((message) => message.id !== id));
-  }, []);
-
+  const { messages, handleRemoveMessage } = useToastContainer();
   return (
     <Container>
       {messages.map((message) => (
