+++ src/pages/NewContact/index.js
@@ -1,20 +1,10 @@
-import React, { useCallback, useRef } from 'react';
+import React from 'react';
 import PageHeader from '../../components/PageHeader';
 import ContactForm from '../../components/ContactForm';
-import ContactsService from '../../services/ContactsService';
-import { toastError, toastSuccess } from '../../utils/toast';
+import useNewContact from './useNewContact';
 
 function NewContact() {
-  const contactForm = useRef(null);
-  const handleSubmit = useCallback(async (contact) => {
-    try {
-      await ContactsService.createContact(contact);
-
-      toastSuccess('Contato cadastrado com sucesso.');
-    } catch (error) {
-      toastError('Ocorreu um erro ao cadastrar o contato');
-    }
-  }, []);
+  const { contactForm, handleSubmit } = useNewContact();
 
   return (
     <>
