+++ src/routes/index.routes.js
@@ -1,7 +1,7 @@
 import { Switch, Route } from 'react-router-dom';
 
 import Home from '../pages/Home';
-import { Container as EditContactPage } from '../pages/EditContact';
+import EditContactPage from '../pages/EditContact';
 import NewContact from '../pages/NewContact';
 
 export default function Routes() {
